refactor: move function
