+++ test/mock-work-dir.ts
@@ -6,6 +6,7 @@ import {
   emptyProjectManifest,
   UnityProjectManifest,
 } from "../src/types/project-manifest";
+import { createProjectVersionTxt } from "../src/utils/project-version-io";
 
 export type ManifestCreationOptions = {
   manifest: boolean | UnityProjectManifest;
@@ -14,16 +15,6 @@ export type ManifestCreationOptions = {
 export const getWorkDir = function (pathToTmp: string): string {
   return path.join(os.tmpdir(), pathToTmp);
 };
-
-export function createProjectVersionTxt(
-  projectDirPath: string,
-  version: string
-) {
-  const projectSettingsDir = path.join(projectDirPath, "ProjectSettings");
-  fse.mkdirpSync(projectSettingsDir);
-  const data = `m_EditorVersion: ${version}`;
-  fse.writeFileSync(path.join(projectSettingsDir, "ProjectVersion.txt"), data);
-}
 
 export const createWorkDir = function (
   pathToTmp: string,
